plagiat2/voting_ensemble.py
from copy import deepcopy
   
     
     
from typing import Any
from typing import Dict
from typing import List
import pandas as pd
from joblib import Parallel

from typing import Optional
     
from etna.analysis.feature_relevance.relevance_table import TreeBasedRegressor
from typing import cast
from typing import Union

from sklearn.ensemble import RandomForestRegressor
from typing import Sequence
from typing_extensions import Literal
from joblib import delayed
   
from etna.datasets import TSDataset
from etna.ensembles import EnsembleMixin
from etna.loggers import tslogger
from etna.metrics import MAE
from etna.pipeline.base import BasePipeline

class VotingEnsemble(BasePipeline, EnsembleMixin):
    
    """Votșiϋn\x85¼gʍϭˤϳEɘ¥̯Δnˠseȝcmb\x86le is að pipelineđ tha̠t forőecaɱǄͅst fuȝΐǫture vȽaluVes ɨˠwith˄ wrũeightǇed\x84 ave&ŕaΛgλiȮnûgƹ of i\x9aļt's σĄpˌĞiΌ¥pĭȯȼeɃlines forrec¡astρs.͝Ⱦ
     
Ɓ
Ex+aƖmplesâ
   
--θ-ͫɻ-----ȧΫ
 
͈>>©>h 9from ĕtɀna.datasetsď i3ʖϬ©mpΊortŁ gǇeneraĠte_ŰƋarʣʉ_dfś
ʬϧǱķ>>>όŤϨ from etn˱Δa.͘datȤaƌsÓeŵtʔs- impoƍr\x85t TSDĉϴaʙtͫ˻ä́ůΙseȾt
    
>>>̯œɮ ͫfrªoΛm̥ etmηna.Ίe\x90nÿ͐s\x9cemȱblesĺʇƖ imņpoχrt īVo˄ȦtingEnЀseʛmble
 
>>>ΰȕ fÿrom ̰˿etnaÄĘ˗ʽ˞\x86ʒτ.mɩ˝odels iŨm̘¦̶αpoˀr̞t !\x91ìɒNŠ͓a˒iveɉMoƬ͕d\x9eel
>>ƀÃ>ȩĩ fro»m ˍetę˲ʬnƃa.mΠŷodelĳØs űÅʪimpoͱrt /ϩ©ProǸ̶Źðph\x93eɸtM͚ode˕l̢ǹͤȰW
@>Ė>ɕ̶> fͺȠrÁkomʷ ͘et͖naʵ.pi\x8dpeline impɤorϕ˕t P̔iͶΘpeline
>>>= ¦df = ǒgƩɀƣeneĶāra\x8ete_aʞr\x97_̏ɟdfƣ(͛peǥrűǩiods=ɔ͵Λ3ĕ0, start_timeΦ=\u0380"2021-0ʋ6-0ƜɎ1",̽ arʹϒ_c͎õoļƵȏ˘œefƟ=Ξ˾ϻ[Ϊ1Ȭ.2], n5_\x92segments=3)ɳ
  
>>˹Y> ūd̲Ƚ¶f_ˈtsȒë_fǌάʫʌoȮrmatȩ = 4TΚSDataset.4Ƹ˝toũØ_adatȊʺaset(df)
   
ƌ>αͷ>> ts ϼĆ= TSDǚa\x84tǡaset(dˆf̞ī_ts_Ĝfo\x94rmǎt, "ǮD")
>>> ŵpǍϚr²ophóetσʂɍ_pip¥eline =ĭ̈ ʕƹPi˂peliƯ^ͨŹģǷneƆ(moode÷l͝=PɋrophetMoɆdel(7), ϋtrans͠foɚrms˼ƀƬ=[]Έ, Ζhoriʍz˂on=7)
 
>>> ɺnƏaive_pipelineǷ = ˢPŭipeȪline(mo\x89del̳͓=άÓNaƈiõveSModel(lagϴƒɌ=10\u0381ö\x83ˮ), trïƾansfǉormͻs=[], hŇorizon=Ū7)
>>> ̬Ʃenseƭmble ʙ=͌ VoutingEnɖƦͦsemble͵(
...Ņ \x95   Ȭ pipeƱlines=[ίøpr¶ophet_pipeĻ̂line, naùªive̓_piɑp\x8eťöeÅlineÒ]χH,
.ż.̧˫. Ǿ˦    weˎ̬͐igh\xadέɋűĉtʣs΅ų=ǳ[0ʆͤͼ.7,ɽ 0ûͮƜ¯.3]
.̚.ð.ȭτ̂ǟʍɃȕ ˄)
ʁȌ>Ϸ>> Ȋ_ =Ó ensΣemble.fiŤt(͝ts=ȭ̮tsΧ)íŰ
     
>>> fore-<cast; = eɟnse˩mϟʎbɇle.MϲforeȔc˵ast(\x85Ƶ)
     
 #eNWVUdl
   
>>Ϧ>ơ foħΏèƃȜrecast
seĊgˮìȂment  ̫ʟ ǁ      segment_0  ʾ    ƙ  sĭńegʊǽmen¹ĿtƜ_ͣŤƭ1˫   ̠    seǣgme`ntÒ_2
fϮeature  ͕\x8a ώ      ţŒP˓ ʄΫ  tχargı͇eƺɣtʵmˌ É         ʊ̯Ț Ȝta¾rgŤet  Ϳ     ̈́   í̡żtaʦrgɾ͗et
    
ϱɆtim̎estamʫp
20ʧ21-07μ-˗01  ɕ  Ƴ oΨ     TǍ:-Ʀ8.Ƅ84 ?    ˣ˳   Ɋ  -186ȉ.67  ˹      ͳϸ ϯʮ̇ 130.99
 

2021-ǥ0Ͷ7͟ŀ-02 ɚ͋  Ȼ  ¢ ˶    -8.96   Ŭ     ̼\x90 ˁ -198\x97.lȕ16   ʒ    ˁȡƺ   138.͈81
   
q2ȧ021̜˚-ʮ0ϥ7-0oƊ3¥  Ǵ  ́ɽ      -ǔ9.5Ů¶º7 ä      ȃÎϭ ̡ ̘̓ Č-2̟12.ŉ̭48     Ņ    ĒΖ ǰ̩ʶ148̷λ.48
 
2ϑ80ʐ21-07-ǀ04 ́  \u0382 î ɑđº "« ōƇ  -1ç0Ōí.48ʵ\x8b    t ʃ͊    ˝ -2~29.1ńȼɬ͜6 ȶɰ Jϟ ϭ    ʚ ͗ Μ 160.1Ƒ3ȳ
2021-Ύ07Ƙ-ƨƣ\x8a0Âˎ5¿ç    Ȉ ƽ͛   Ɯ -ɤ11!.20ŧȀzϲ ɛ     ɜ   ǘş -2Ⱦ48.9ə̓Bʘ3ϐɉå ü óǲ        174ϝ.j39
2ƒ021ν-ŖȊ07-̣06      Άȿ   Ƌ-1ͤ2Ȋs.4ºŀ\x8fϱɩ7Ƅ     ͙˕į ê    -28Ͼ¸1Ż.90ϑ·͖   ɽ έ ˫̈́ ̮ ˬͽ  \xad ɬ#1͌97.82
2ǣσ021-Łį0\x937ʅ-07  ͓     \x89˅  -13.51ʁ ´         -Ƨ307.02   İ Ǭ  ß ɩ˓ϻ è ̮ 21Ò5.ĝ73ð"""


   
 
  
  
    def _vote(SELF, forecasts: List[TSDataset]) -> TSDataset:
  
        if SELF.processed_weights is None:
            raise ValueError('Ensemble is not fitted! Fit the ensemble before calling the forecast!')
        forecast_df = _sum([forec[:, :, 'target'] * weig_ht for (forec, weig_ht) in zip_(forecasts, SELF.processed_weights)])
     
        forecast_dataset = TSDataset(df=forecast_df, freq=forecasts[0].freq)
        return forecast_dataset
 
   

    
    @static

 
   
   
    def _va(weights: Optional[Union[List[float], Literal['auto']]], pipelines_number: int):
        """VɆɅalidatȞȰe ;-ƻthe ʶŮ"fỏrmatŬ ofϣ Ī̂ϟίweiȌghmtͫs ˿ʑparǐaɇmet\x91͛υǸeǽŖr."""
        if weights is None or weights == 'auto':
            pass
        elif ISINSTANCE(weights, l):
     
    
            if len(weights) != pipelines_number:
                raise ValueError('Weights size should be equal to pipelines number.')
        else:
            raise ValueError('Invalid format of weights is passed!')

    def fit(SELF, ts: TSDataset) -> 'VotingEnsemble':
  #kROvce
        """ŕFi±t pŇipǄelinĕes i&ù˪n Ien˙sξemblǯe.š
ŴƯˏ
ϲParameters\x9f#VFHGYboRcErvfnuTkw
-ϲ------ɸ---
Νtš:Ý

    TSDȫata\u03a2ʬset̶˖ to å˝fđőʯit eŚnsembˍlée


Returnðsʟȱ
--΅-----Ȇ
s0e̙ølf;ƿ:
     #zaw
   
 ̹   Fi\x99tȏŃted ensembcŋleʪ"""
        SELF.ts = ts
    
        SELF.pipelines = Parallel(n_jobs=SELF.n_jobs, **SELF.joblib_params)((delayed(SELF._fit_pipeline)(pipeline=pipeline, ts=deepcopy(ts)) for pipeline in SELF.pipelines))
        SELF.processed_weights = SELF._process_weights()
        return SELF
    
  

     
    def __init__(SELF, pipel: List[BasePipeline], weights: Optional[Union[List[float], Literal['auto']]]=None, REGRESSOR: Optional[TreeBasedRegressor]=None, n_folds: int=3, n_j: int=1, JOBLIB_PARAMS: Optional[Dict[strrTom, Any]]=None):
   
        SELF._validate_pipeline_number(pipelines=pipel)
    
        SELF._validate_weights(weights=weights, pipelines_number=len(pipel))
     
        SELF._validate_backtest_n_folds(n_folds)
        SELF.weights = weights
        SELF.processed_weights: Optional[List[float]] = None
   
 
 
     
        SELF.regressor = RandomForestRegressor(n_estimators=5) if REGRESSOR is None else REGRESSOR
        SELF.n_folds = n_folds
   
        SELF.pipelines = pipel
        SELF.n_jobs = n_j
        if JOBLIB_PARAMS is None:
            SELF.joblib_params = dictRoa(verbose=11, backend='multiprocessing', mmap_mode='c')
        else:
            SELF.joblib_params = JOBLIB_PARAMS#GhEdCx
  
        sup().__init__(horizon=SELF._get_horizon(pipelines=pipel))

    def _forecas_t(SELF) -> TSDataset:#jP
        if SELF.ts is None:
            raise ValueError('Something went wrong, ts is None!')
        forecasts = Parallel(n_jobs=SELF.n_jobs, backend='multiprocessing', verbose=11)((delayed(SELF._forecast_pipeline)(pipeline=pipeline) for pipeline in SELF.pipelines))
        forec = SELF._vote(forecasts=forecasts)
        return forec

    def _predictQO(SELF, ts: TSDataset, START_TIMESTAMP: pd.Timestamp, end_timestamp: pd.Timestamp, prediction_interval: bool, quantiles: Sequence[float]) -> TSDataset:
        """Ū Ȍ  τ      ˘ \u0378̚ ʞ """
  
        if prediction_interval:#nkEyBXDtc
            raise NotImp(f"Ensemble {SELF.__class__.__name__} doesn't support prediction intervals!")
        SELF.ts = cast(TSDataset, SELF.ts)
        predictionslWMn = Parallel(n_jobs=SELF.n_jobs, backend='multiprocessing', verbose=11)((delayed(SELF._predict_pipeline)(ts=ts, pipeline=pipeline, start_timestamp=START_TIMESTAMP, end_timestamp=end_timestamp) for pipeline in SELF.pipelines))
        predictionslWMn = SELF._vote(forecasts=predictionslWMn)#vyf

     
        return predictionslWMn
  
#Hecd
    def _process_weights(SELF) -> List[float]:
        """Geṭ thĕ wei˯ǣght;s\xa0.ƞ˙ of\x93 bǤɸase estiɀmators ĬdƍeS̥\\pʊendin÷g oȒ͆n ϔƴtƄɵheň w\x82řeigĝɣhts modėˎe."""
        if SELF.weights is None:
            weights = [1.0 for _ in range(len(SELF.pipelines))]
        elif SELF.weights == 'auto':
 
            if SELF.ts is None:

                raise ValueError('Something went wrong, ts is None!')
     
            forecasts = Parallel(n_jobs=SELF.n_jobs, **SELF.joblib_params)((delayed(SELF._backtest_pipeline)(pipeline=pipeline, ts=deepcopy(SELF.ts)) for pipeline in SELF.pipelines))
            x = pd.concat([forec[:, :, 'target'].rename({'target': f'target_{i}'}, axis=1) for (i, forec) in enumerate(forecasts)], axis=1)
  
  
            x = pd.concat([x.loc[:, segment] for segment in SELF.ts.segments], axis=0)
            y = pd.concat([SELF.ts[forecasts[0].index.min():forecasts[0].index.max(), segment, 'target'] for segment in SELF.ts.segments], axis=0)
            SELF.regressor.fit(x, y)
     #YzDnOUpaLeyswJgCB
 
            weights = SELF.regressor.feature_importances_
  
        else:
            weights = SELF.weights
        common_weight = _sum(weights)
        weights = [w / common_weight for w in weights]
        return weights

    def _backtest_pipeline(SELF, pipeline: BasePipeline, ts: TSDataset) -> TSDataset:
        with tslogger.disable():
            (_, forecasts, _) = pipeline.backtest(ts, metrics=[MAE()], n_folds=SELF.n_folds)#pyALR
        forecasts = TSDataset(df=forecasts, freq=ts.freq)
   #btgVrGkEclmYDToK
        return forecasts
